pyproject.toml
[project]
name = "dtse-test-task"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "django>=5.2.4",
    "django-rest-knox>=5.0.2",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "joblib==1.2.0",
    "numpy==1.23.4",
    "pandas==1.5.0",
    "python-decouple>=3.8",
    "scikit-learn==1.1.2",
]

[dependency-groups]
prod = [
    "gunicorn>=23.0.0",
]
